chore: move RandomAddressBookGenerator
