+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/DummyHashgraph.java
@@ -17,8 +17,8 @@
 package com.swirlds.platform;
 
 import com.swirlds.common.platform.NodeId;
-import com.swirlds.common.test.fixtures.RandomAddressBookGenerator;
 import com.swirlds.platform.system.address.AddressBook;
+import com.swirlds.platform.test.fixtures.addressbook.RandomAddressBookGenerator;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.HashMap;
 import java.util.Random;
